+++ hwDraftComments_Модульный_JS/api.js
@@ -11,6 +11,14 @@ export function getComments () {
     }
      return response.json();
     })
+  .catch((error) => {
+    if (error.message === "Сервер упал") {
+      alert("Нет интернета");
+    }
+    if (error.message === "Failed to fetch") {
+      alert("Кажется что-то пошло не так, попробуй позже..");
+    }
+  });
 };
 export function getPost ({name, text}) {
   return fetch("https://wedev-api.sky.pro/api/v1/:anna-kalina/comments", {
@@ -19,14 +27,25 @@ export function getPost ({name, text}) {
        name: name,
        text: text,
      }),
-    })
-    .then((response) => {
-     if (response.status === 500) {
-       throw new Error("Сервер упал");
-     }
-     if (response.status === 400) {
-       throw new Error("Вводимые данные слишком короткие");
-     }  
-      return response.json();
-    }) 
+  })
+  .then((response) => {
+    if (response.status === 500) {
+      throw new Error("Сервер упал");
+    }
+    if (response.status === 400) {
+      throw new Error("Вводимые данные слишком короткие");
+    }  
+    return response.json();
+  })
+  .catch((error) => {
+    if (error.message === "Сервер упал") {
+      alert("Нет интернета");
+    }
+    if (error.message === "Вводимые данные слишком короткие") {
+      alert("Имя или текст менее трех символов");
+    }
+    if (error.message === "Failed to fetch") {
+      alert("Кажется что-то пошло не так, попробуй позже..");
+    }
+  })   
 };
