+++ hwDraftComments_Модульный_JS/main.js
@@ -1,20 +1,17 @@
-import { getComments } from "./api.js";
-import { getPost } from "./api.js";
-import { renderCommentators } from "./renderCommentators.js";
+import { getComments, getPost  } from "./api.js";
+import { сommentators, renderCommentators } from "./renderCommentators.js";
 
 
 window.onload = function() {
   let preloader = document.getElementById('preloader');
 };
 
-const listElement = document.getElementById ('list');
 const nameElement = document.getElementById ('name-input');
 const textElement = document.getElementById ('text-input');
 const buttonElement = document.getElementById ('add-button');
 const addForm = document.getElementById("form");
 const loader = document.querySelector(".loader");
 
-let сommentators = [];
 
 const getCom = () => {
 
@@ -26,46 +23,20 @@ const getCom = () => {
         comment : comment.text,
         likes : comment.likes,
         isLiked: comment.isLiked,
+        forceError: true,
       };
     });
 
     сommentators = appComments;
-    renderCommentators({сommentators, answerComment, editEventListeners, initEventListeners});
+    renderCommentators();
     loader.textContent = '';
     addForm.classList.remove("hidden");
     preloader.classList.add('preloader-hidden');
       
   })
-  .catch((error) => {
-    if (error.message === "Сервер упал") {
-      alert("Нет интернета");
-    }
-    if (error.message === "Failed to fetch") {
-      alert("Кажется что-то пошло не так, попробуй позже..");
-    }
-  });  
-
 };
+
 getCom();
-
-function answerComment () {
-  
-  const commentsElements = document.querySelectorAll('.comment');
-
-  for (const commentsEl of commentsElements) { 
-
-    commentsEl.addEventListener('click', (e) => {
-      e.stopPropagation();
-
-      const index = commentsEl .dataset.index;
-    
-      textElement.value = `QUOTE_BEGIN${сommentators[index].comment}\n${сommentators[index].name}QUOTE_END`;
-
-      renderCommentators({сommentators, answerComment, editEventListeners, initEventListeners});
-
-    })
-  }
-};
 
 const editEventListeners = () => { 
   const editButtonElements = document.querySelectorAll(".edit-button"); 
@@ -74,7 +45,7 @@ const editEventListeners = () => {
       e.stopPropagation();
       const index = editButtonEl.dataset.index;
       сommentators[index].isEdit = !сommentators[index].isEdit;
-      renderCommentators({сommentators, answerComment, editEventListeners, initEventListeners});
+      renderCommentators();
     });
   }
 
@@ -85,55 +56,10 @@ const editEventListeners = () => {
       const index = saveButton.dataset.index;
       сommentators[index].isEdit = !сommentators[index].isEdit;
       сommentators[index].comment = saveButton.closest('.comment').querySelector('textarea').value
-      renderCommentators({сommentators, answerComment, editEventListeners, initEventListeners});
+      renderCommentators();
     });
   }
 }
-
-function delay(interval = 300) {
-  return new Promise((resolve) => {
-    setTimeout(() => {
-      resolve();
-    }, interval);
-  });
-}   
-
-const initEventListeners = () => {
-  const likeButtonElement = document.querySelectorAll(".like-button");
-  for ( const likeButtonEl of likeButtonElement) {
-
-    const index = likeButtonEl.dataset.index;
-    const counter = likeButtonEl.dataset.like;
-
-    likeButtonEl.addEventListener("click", (event) => {
-      event.stopPropagation();
-
-      likeButtonEl.classList.add("-loading-like");
-
-      delay(2000).then(() => {
-
-        if (сommentators[index].isLiked === false) {
-
-          const result = Number(counter) + 1;
-
-          сommentators[index].likes = result;
-          сommentators[index].isLiked = true;
-
-        } else if (сommentators[index].isLiked === true) {
-
-          const result = Number(counter) - 1;
-
-          сommentators[index].likes = result;
-          сommentators[index].isLiked = false;
-        }
-
-        renderCommentators({сommentators, answerComment, editEventListeners, initEventListeners});
-      })  
-    })
-  }
-}
-renderCommentators({сommentators, answerComment, editEventListeners, initEventListeners});
-
 
 buttonElement.addEventListener("click", () => {
   nameElement.classList.remove("error");
@@ -161,6 +87,15 @@ buttonElement.addEventListener("click", () => {
    return `${date}.${month}.${year} ${h}:${m}`
   }; 
 
+  const sanitizeHtml = (htmlString) => {
+    return htmlString
+    .replaceAll("&", "&amp;")
+    .replaceAll("<", "&lt;")
+    .replaceAll(">", "&gt;")
+    .replaceAll('"', "&quot;")
+    .replaceAll("QUOTE_BEGIN", "<div class='quote'>")
+    .replaceAll("QUOTE_END", "</div>")
+  };
 
   addForm.classList.add("hidden");
   loader.textContent = 'Комментарий добавляется .....';
@@ -170,31 +105,16 @@ buttonElement.addEventListener("click", () => {
     text: textElement.value,
   })
   .then(() => {
-      nameElement.value = "";
-      textElement.value = "";
-      return getCom();
-    })
-    .catch((error) => {
-       
-      if (error.message === "Сервер упал") {
-        alert("Нет интернета");
-
-      }
-      if (error.message === "Вводимые данные слишком короткие") {
-        alert("Имя или текст менее трех символов");
-
-      }
-      if (error.message === "Failed to fetch") {
-        alert("Кажется что-то пошло не так, попробуй позже..");
-      }
-      console.warn(error);
-      loader.textContent = '';
-      addForm.classList.remove("hidden");
-    });
-
-    renderCommentators({сommentators, answerComment, editEventListeners, initEventListeners});
+    nameElement.value = "";
+    textElement.value = "";
+    return getCom();
+  })
+  console.warn(error);
+  loader.textContent = '';
+  addForm.classList.remove("hidden");
 });
-renderCommentators({сommentators, answerComment, editEventListeners, initEventListeners})
+
+renderCommentators();
 
 const formElement = document.getElementById ('form');
  formElement.addEventListener('keyup', keyEvent);
@@ -207,7 +127,7 @@ const formElement = document.getElementById ('form');
 const removeButton = document.getElementById("deleteComment");
 removeButton.addEventListener("click", () => {
   сommentators.pop();
-  renderCommentators({сommentators, answerComment, editEventListeners, initEventListeners});
+  renderCommentators();
 });
 
 console.log("It works!");
